+++ contracts/cw721-base/examples/schema.rs
@@ -1,12 +1,13 @@
 use std::env::current_dir;
 use std::fs::create_dir_all;
 
-use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
+use cosmwasm_schema::{export_schema, export_schema_with_title, remove_schemas, schema_for};
 
 use cw721::{
     AllNftInfoResponse, ApprovedForAllResponse, ContractInfoResponse, NftInfoResponse,
     NumTokensResponse, OwnerOfResponse, TokensResponse,
 };
+use cw721_base::entry::Extension;
 use cw721_base::msg::{ExecuteMsg, InstantiateMsg, MinterResponse, QueryMsg};
 
 fn main() {
@@ -16,13 +17,21 @@ fn main() {
     remove_schemas(&out_dir).unwrap();
 
     export_schema(&schema_for!(InstantiateMsg), &out_dir);
-    export_schema(&schema_for!(ExecuteMsg), &out_dir);
+    export_schema_with_title(&schema_for!(ExecuteMsg<Extension>), &out_dir, "ExecuteMsg");
     export_schema(&schema_for!(QueryMsg), &out_dir);
-    export_schema(&schema_for!(AllNftInfoResponse), &out_dir);
+    export_schema_with_title(
+        &schema_for!(AllNftInfoResponse<Extension>),
+        &out_dir,
+        "AllNftInfoResponse",
+    );
     export_schema(&schema_for!(ApprovedForAllResponse), &out_dir);
     export_schema(&schema_for!(ContractInfoResponse), &out_dir);
     export_schema(&schema_for!(MinterResponse), &out_dir);
-    export_schema(&schema_for!(NftInfoResponse), &out_dir);
+    export_schema_with_title(
+        &schema_for!(NftInfoResponse<Extension>),
+        &out_dir,
+        "NftInfoResponse",
+    );
     export_schema(&schema_for!(NumTokensResponse), &out_dir);
     export_schema(&schema_for!(OwnerOfResponse), &out_dir);
     export_schema(&schema_for!(TokensResponse), &out_dir);
