+++ contracts/cw721-base/schema/all_nft_info_response.json
@@ -19,7 +19,7 @@
       "description": "Data on the token itself,",
       "allOf": [
         {
-          "$ref": "#/definitions/NftInfoResponse"
+          "$ref": "#/definitions/NftInfoResponse_for_Nullable_Empty"
         }
       ]
     }
@@ -45,6 +45,10 @@
           "type": "string"
         }
       }
+    },
+    "Empty": {
+      "description": "An empty struct that serves as a placeholder in different places, such as contracts that don't set a custom message.\n\nIt is designed to be expressable in correct JSON and JSON Schema but contains no meaningful data. Previously we used enums without cases, but those cannot represented as valid JSON Schema (https://github.com/CosmWasm/cosmwasm/issues/451)",
+      "type": "object"
     },
     "Expiration": {
       "description": "Expiration represents a point in time when some event happens. It can compare with a BlockInfo and will return is_expired() == true once the condition is hit (and for every block in the future)",
@@ -92,7 +96,7 @@
         }
       ]
     },
-    "NftInfoResponse": {
+    "NftInfoResponse_for_Nullable_Empty": {
       "type": "object",
       "required": [
         "description",
@@ -102,6 +106,17 @@
         "description": {
           "description": "Describes the asset to which this NFT represents",
           "type": "string"
+        },
+        "extension": {
+          "description": "You can add any custom metadata here when you extend cw721-base",
+          "anyOf": [
+            {
+              "$ref": "#/definitions/Empty"
+            },
+            {
+              "type": "null"
+            }
+          ]
         },
         "image": {
           "description": "\"A URI pointing to a resource with mime type image/* representing the asset to which this NFT represents. Consider making any images at a width between 320 and 1080 pixels and aspect ratio between 1.91:1 and 4:5 inclusive. TODO: Use https://docs.rs/url_serde for type-safety",
