+++ contracts/cw721-base/schema/execute_msg.json
@@ -177,7 +177,7 @@
       ],
       "properties": {
         "mint": {
-          "$ref": "#/definitions/MintMsg"
+          "$ref": "#/definitions/MintMsg_for_Nullable_Empty"
         }
       },
       "additionalProperties": false
@@ -187,6 +187,10 @@
     "Binary": {
       "description": "Binary is a wrapper around Vec<u8> to add base64 de/serialization with serde. It also adds some helper methods to help encode inline.\n\nThis is only needed as serde-json-{core,wasm} has a horrible encoding for Vec<u8>",
       "type": "string"
+    },
+    "Empty": {
+      "description": "An empty struct that serves as a placeholder in different places, such as contracts that don't set a custom message.\n\nIt is designed to be expressable in correct JSON and JSON Schema but contains no meaningful data. Previously we used enums without cases, but those cannot represented as valid JSON Schema (https://github.com/CosmWasm/cosmwasm/issues/451)",
+      "type": "object"
     },
     "Expiration": {
       "description": "Expiration represents a point in time when some event happens. It can compare with a BlockInfo and will return is_expired() == true once the condition is hit (and for every block in the future)",
@@ -234,7 +238,7 @@
         }
       ]
     },
-    "MintMsg": {
+    "MintMsg_for_Nullable_Empty": {
       "type": "object",
       "required": [
         "name",
@@ -247,6 +251,17 @@
           "type": [
             "string",
             "null"
+          ]
+        },
+        "extension": {
+          "description": "Any custom extension used by this contract",
+          "anyOf": [
+            {
+              "$ref": "#/definitions/Empty"
+            },
+            {
+              "type": "null"
+            }
           ]
         },
         "image": {
