+++ packages/cw721/examples/schema.rs
@@ -1,13 +1,16 @@
 use std::env::current_dir;
 use std::fs::create_dir_all;
 
-use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
+use cosmwasm_schema::{export_schema, export_schema_with_title, remove_schemas, schema_for};
+use cosmwasm_std::Empty;
 
 use cw721::{
     AllNftInfoResponse, ApprovedForAllResponse, ContractInfoResponse, Cw721ExecuteMsg,
     Cw721QueryMsg, Cw721ReceiveMsg, NftInfoResponse, NumTokensResponse, OwnerOfResponse,
     TokensResponse,
 };
+
+type Extension = Option<Empty>;
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
@@ -18,11 +21,19 @@ fn main() {
     export_schema(&schema_for!(Cw721ExecuteMsg), &out_dir);
     export_schema(&schema_for!(Cw721QueryMsg), &out_dir);
     export_schema(&schema_for!(Cw721ReceiveMsg), &out_dir);
-    export_schema(&schema_for!(AllNftInfoResponse), &out_dir);
+    export_schema_with_title(
+        &schema_for!(AllNftInfoResponse<Extension>),
+        &out_dir,
+        "AllNftInfoResponse",
+    );
     export_schema(&schema_for!(ApprovedForAllResponse), &out_dir);
     export_schema(&schema_for!(ContractInfoResponse), &out_dir);
     export_schema(&schema_for!(OwnerOfResponse), &out_dir);
-    export_schema(&schema_for!(NftInfoResponse), &out_dir);
+    export_schema_with_title(
+        &schema_for!(NftInfoResponse<Extension>),
+        &out_dir,
+        "NftInfoResponse",
+    );
     export_schema(&schema_for!(NumTokensResponse), &out_dir);
     export_schema(&schema_for!(TokensResponse), &out_dir);
 }
