+++ packages/cw721/schema/nft_info_response.json
@@ -11,6 +11,17 @@
       "description": "Describes the asset to which this NFT represents",
       "type": "string"
     },
+    "extension": {
+      "description": "You can add any custom metadata here when you extend cw721-base",
+      "anyOf": [
+        {
+          "$ref": "#/definitions/Empty"
+        },
+        {
+          "type": "null"
+        }
+      ]
+    },
     "image": {
       "description": "\"A URI pointing to a resource with mime type image/* representing the asset to which this NFT represents. Consider making any images at a width between 320 and 1080 pixels and aspect ratio between 1.91:1 and 4:5 inclusive. TODO: Use https://docs.rs/url_serde for type-safety",
       "type": [
@@ -21,6 +32,12 @@
     "name": {
       "description": "Identifies the asset to which this NFT represents",
       "type": "string"
+    }
+  },
+  "definitions": {
+    "Empty": {
+      "description": "An empty struct that serves as a placeholder in different places, such as contracts that don't set a custom message.\n\nIt is designed to be expressable in correct JSON and JSON Schema but contains no meaningful data. Previously we used enums without cases, but those cannot represented as valid JSON Schema (https://github.com/CosmWasm/cosmwasm/issues/451)",
+      "type": "object"
     }
   }
 }
