+++ packages/cw721/src/traits.rs
@@ -1,0 +1,140 @@
+use schemars::JsonSchema;
+use serde::de::DeserializeOwned;
+use serde::Serialize;
+
+use crate::{
+    AllNftInfoResponse, ApprovedForAllResponse, ContractInfoResponse, NftInfoResponse,
+    NumTokensResponse, OwnerOfResponse, TokensResponse,
+};
+use cosmwasm_std::{Binary, Deps, DepsMut, Empty, Env, MessageInfo, Response, StdResult};
+use cw0::Expiration;
+
+// TODO: move this somewhere else... ideally cosmwasm-std
+pub trait CustomMsg: Clone + std::fmt::Debug + PartialEq + JsonSchema {}
+
+impl CustomMsg for Empty {}
+
+pub trait Cw721<T, C>: Cw721Execute<T, C> + Cw721Query<T>
+where
+    T: Serialize + DeserializeOwned + Clone,
+    C: CustomMsg,
+{
+}
+
+pub trait Cw721Execute<T, C>
+where
+    T: Serialize + DeserializeOwned + Clone,
+    C: CustomMsg,
+{
+    type Err: ToString;
+
+    fn transfer_nft(
+        &self,
+        deps: DepsMut,
+        env: Env,
+        info: MessageInfo,
+        recipient: String,
+        token_id: String,
+    ) -> Result<Response<C>, Self::Err>;
+
+    fn send_nft(
+        &self,
+        deps: DepsMut,
+        env: Env,
+        info: MessageInfo,
+        contract: String,
+        token_id: String,
+        msg: Binary,
+    ) -> Result<Response<C>, Self::Err>;
+
+    fn approve(
+        &self,
+        deps: DepsMut,
+        env: Env,
+        info: MessageInfo,
+        spender: String,
+        token_id: String,
+        expires: Option<Expiration>,
+    ) -> Result<Response<C>, Self::Err>;
+
+    fn revoke(
+        &self,
+        deps: DepsMut,
+        env: Env,
+        info: MessageInfo,
+        spender: String,
+        token_id: String,
+    ) -> Result<Response<C>, Self::Err>;
+
+    fn approve_all(
+        &self,
+        deps: DepsMut,
+        env: Env,
+        info: MessageInfo,
+        operator: String,
+        expires: Option<Expiration>,
+    ) -> Result<Response<C>, Self::Err>;
+
+    fn revoke_all(
+        &self,
+        deps: DepsMut,
+        env: Env,
+        info: MessageInfo,
+        operator: String,
+    ) -> Result<Response<C>, Self::Err>;
+}
+
+pub trait Cw721Query<T>
+where
+    T: Serialize + DeserializeOwned + Clone,
+{
+    // TODO: use custom error?
+    // How to handle the two derived error types?
+
+    fn contract_info(&self, deps: Deps) -> StdResult<ContractInfoResponse>;
+
+    fn num_tokens(&self, deps: Deps) -> StdResult<NumTokensResponse>;
+
+    fn nft_info(&self, deps: Deps, token_id: String) -> StdResult<NftInfoResponse<T>>;
+
+    fn owner_of(
+        &self,
+        deps: Deps,
+        env: Env,
+        token_id: String,
+        include_expired: bool,
+    ) -> StdResult<OwnerOfResponse>;
+
+    fn all_approvals(
+        &self,
+        deps: Deps,
+        env: Env,
+        owner: String,
+        include_expired: bool,
+        start_after: Option<String>,
+        limit: Option<u32>,
+    ) -> StdResult<ApprovedForAllResponse>;
+
+    fn tokens(
+        &self,
+        deps: Deps,
+        owner: String,
+        start_after: Option<String>,
+        limit: Option<u32>,
+    ) -> StdResult<TokensResponse>;
+
+    fn all_tokens(
+        &self,
+        deps: Deps,
+        start_after: Option<String>,
+        limit: Option<u32>,
+    ) -> StdResult<TokensResponse>;
+
+    fn all_nft_info(
+        &self,
+        deps: Deps,
+        env: Env,
+        token_id: String,
+        include_expired: bool,
+    ) -> StdResult<AllNftInfoResponse<T>>;
+}
